Move encoder classes
